feat(revenue) : DIVIDE - Revenue and Score class
