+++ src/main/kotlin/blackjack/Score.kt
@@ -1,0 +1,7 @@
+package blackjack
+
+data class Score(
+    val isWin: Boolean = false,
+    val isDraw: Boolean = false,
+    val isBlackJack: Boolean = false
+)
